move passthrough tokenizer to its own file
